rename eslint-rules directory to lint-rules
