+++ eslint-plugin-epic-web.js
@@ -1,2 +1,2 @@
-export { default } from './eslint-rules/eslint-plugin-epic-web.js'
-export { preferDisposeInTestsRule } from './eslint-rules/eslint-plugin-epic-web.js'
+export { default } from './lint-rules/eslint-plugin-epic-web.js'
+export { preferDisposeInTestsRule } from './lint-rules/eslint-plugin-epic-web.js'
